requirement.txt
Django
EasyProcess
Pillow
PyAutoGUI
PyExecJS
PyMsgBox
PyMySQL
PyMysqlDB
PyScreeze
PyTweening
PyVirtualDisplay
XlsxWriter
action
aircv
asn1crypto
baidu-aip
certifi
cffi
chardet
comtypes
console
cryptography
easygui
et-xmlfile
ezenv
get
idna
jdcal
mysqlclient
numpy
opencv-python
openpyxl
out
pandas
pip
post
psutil
public
pycparser
python-dateutil
pytz
pywin32
pywinauto
pyxvfb
query-string
replace
request
requests
selenium
setuptools
six
urllib3
utils
xlrd
xlutils
xlwings
xlwt
xvfbwrapper